+++ models/document_revision.py
@@ -25,6 +25,8 @@ class DocumentRevision(models.Model):
         is_current_revision = self.document_id.current_revision_id == self
         if is_current_revision:
             raise UserError(_('You cannot delete a Document Revision as long it is set as Current Revision.'))
+        if not self.env.user.has_group('certificate_planer.group_certificate_planer_administrator') and len(self) > 1:
+            raise UserError(_('You cannot delete multiple documents.'))
         return super(DocumentRevision, self).unlink()
 
     def name_get(self):
